pyproject.toml
[tool.poetry]
name = "ezra-klein-books"
version = "0.1.0"
description = ""
authors = ["Zach Bellay <zachbellay@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.9"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
gunicorn = "^21.2.0"
python-decouple = "^3.8"
redis = "^5.0.1"
whitenoise = "^6.6.0"
psycopg = {extras = ["binary"], version = "^3.1.17"}
django-shortuuidfield = "^0.1.3"
feedparser = "^6.0.11"
dateparser = "^1.2.0"
bleach = "^6.1.0"
mutagen = "^1.47.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
djlint = "^1.34.1"
ruff = "^0.1.14"
mypy-extensions = "^1.0.0"
coverage = "^7.4.0"
editorconfig = "^0.12.3"
watchdog = "^3.0.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
